SFS TOC polishing
Fixes staffanm/lagen.nu#5, staffanm/lagen.nu#6, staffanm/lagen.nu#7

This introduces a significant change in util.c_locale (now
switch_locale) and the sorting logic in the toc function, which now
respects a docrepo-specified locale to use for collation. This is
neccesary because the default sv_SE.UTF-8 does not sort swedish
characters in the expected manner.

It also overrides facet_query with a hardcoded SPARQL query, that
knows about the relationship between a consolidated document and the
document its consolidating. This query might need further tweaking
when subjected to large-scale data.