replaced the http_handle function with a requesthandler class
The class is intented to be easier to override in parts and possibly
handle more things related to path/uri translation (eg the
uritransform stuff). In particular, the class needs to be extended to
properly support the repo/dir/attachment parameters used by the
prov:wasDerivedFrom URIs (and which today is handled within the
uritransform function when generating static pages, but isn't handled
at all by the wsgi stuff.

The change was needed in order to support WSGI serving of css and
background images using ?attachment=index.css URIs. Also because it
never felt right in the main DocumentRepository class.