On-demand loading of facsimile pages rendered from the original PDF
If RequestHandler recieves page and format properties in the provided
URI, it will use ImageMagick convert to render the specified page in
the specified format, then return the path to that image.

PDFReader now records such an URI (in the src attribute) for every
page. When rendering to XHTML, this is converted to an <img
data-src="...">

Finally some jquery magic hooked up to a button (in forarbete.xsl) for
every page hides the HTML rendering for that page and on-demand loads
a facsimile image instead.