DV now handles mixed-content metadata for Lagrum, Rättsfall and Litteratur
This means that instead of these types of metadata only being
represented by a list of URIs, a regular (X)HTML fragment containing
links and plaintext is embedded within the BODY of the XHTML doc. No
metadata of these types are stored in the HEAD any longer. dv.xsl is
adapted to use this embedded body metadata.

(This also means that a number of dcterms:identifier triples for
rpubl:rattsfallshanvisning references are removed, as the identifier
is embedded in the body as a regular link, with the identifier as the
link text)